made every question have one equals correct answer, two equals mistyped result and one easily verified upper or lower range, msd or lsd. Removed duplicates

+++ Format.java
@@ -100,7 +100,7 @@ public class Format {
         }
         int digs = absActual > 0? 1 + (int)Math.log10(absActual) : 1;
         long tmp3 = absActual;
-        int ten2pow = (int)Math.pow(10, digs-1);
+        long ten2pow = (long)Math.pow(10, digs-1);
         if( dp >= digs ) {
             actual.append("0.");
         }
